Undo changes to train.py defaults

+++ sockeye/train.py
@@ -402,7 +402,7 @@ def create_encoder_config(args: argparse.Namespace,
         max_seq_len_source=max_seq_len_source,
         max_seq_len_target=max_seq_len_target,
         lhuc=args.lhuc is not None and (C.LHUC_ENCODER in args.lhuc or C.LHUC_ALL in args.lhuc),
-        dtype='int8')
+        dtype='float32')
     encoder_num_hidden = encoder_transformer_model_size
 
     return config_encoder, encoder_num_hidden
@@ -440,7 +440,7 @@ def create_decoder_config(args: argparse.Namespace, encoder_num_hidden: int,
         max_seq_len_source=max_seq_len_source,
         max_seq_len_target=max_seq_len_target,
         lhuc=args.lhuc is not None and (C.LHUC_DECODER in args.lhuc or C.LHUC_ALL in args.lhuc),
-        dtype='int8')
+        dtype='float32')
 
     return config_decoder
